Split ReportPanic across metrics and svclog
